refactor: ensure dagger has correct packages
Dagger had some incorrect packages from everything being moved over
