+++ app/build.gradle
@@ -57,6 +57,7 @@ dependencies {
     implementation project(path: ':account')
     implementation project(path: ':app-discover')
     implementation project(path: ':app-profile')
+    implementation project(path: ':app-series')
     implementation project(path: ':app-settings')
     implementation project(path: ':app-timeline')
     implementation project(path: ':core')
