+++ app/src/main/java/com/chesire/malime/injection/androidmodules/FragmentModule.kt
@@ -2,14 +2,14 @@ package com.chesire.malime.injection.androidmodules
 
 import com.chesire.malime.app.discover.DiscoverFragment
 import com.chesire.malime.app.profile.ProfileFragment
+import com.chesire.malime.app.series.detail.SeriesDetailSheetFragment
+import com.chesire.malime.app.series.list.anime.AnimeFragment
+import com.chesire.malime.app.series.list.manga.MangaFragment
+import com.chesire.malime.app.series.search.SearchFragment
 import com.chesire.malime.app.settings.oss.OssFragment
 import com.chesire.malime.app.timeline.TimelineFragment
 import com.chesire.malime.flow.login.details.DetailsFragment
 import com.chesire.malime.flow.login.syncing.SyncingFragment
-import com.chesire.malime.app.series.detail.SeriesDetailSheetFragment
-import com.chesire.malime.app.series.detail.list.anime.AnimeFragment
-import com.chesire.malime.app.series.detail.list.manga.MangaFragment
-import com.chesire.malime.app.series.search.SearchFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -17,7 +17,7 @@ import dagger.android.ContributesAndroidInjector
 @Module
 abstract class FragmentModule {
     @ContributesAndroidInjector
-    abstract fun contributeAnimeFragment(): com.chesire.malime.app.series.detail.list.anime.AnimeFragment
+    abstract fun contributeAnimeFragment(): AnimeFragment
 
     @ContributesAndroidInjector
     abstract fun contributeDetailsFragment(): DetailsFragment
@@ -26,7 +26,7 @@ abstract class FragmentModule {
     abstract fun contributeDiscoverFragment(): DiscoverFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeMangaFragment(): com.chesire.malime.app.series.detail.list.manga.MangaFragment
+    abstract fun contributeMangaFragment(): MangaFragment
 
     @ContributesAndroidInjector
     abstract fun contributeOssFragment(): OssFragment
@@ -36,10 +36,10 @@ abstract class FragmentModule {
 
     @Suppress("FunctionMaxLength")
     @ContributesAndroidInjector
-    abstract fun contributeSeriesDetailSheetFragment(): com.chesire.malime.app.series.detail.SeriesDetailSheetFragment
+    abstract fun contributeSeriesDetailSheetFragment(): SeriesDetailSheetFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeSeriesSearchFragment(): com.chesire.malime.app.series.search.SearchFragment
+    abstract fun contributeSeriesSearchFragment(): SearchFragment
 
     @ContributesAndroidInjector
     abstract fun contributeSyncingFragment(): SyncingFragment
